feat(types): добавлены типы-аннотации для взаимодействия с Access Token репозитория/группы + типы-аннотации для Access Token размещены по файлам
